+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- `init --git` will not initialize a repository unless git has configured both, `name` and `email`
+
 
 ## [2.6.0] - 2020-12-31